refactor: Simplify index.ts. Update tests to match
