+++ src/index.ts
@@ -1,22 +1,4 @@
 // Two basic actions available with v6.0.0:
-
-import { createTranslator } from "./convertor";
-
-import alphabets from "./constants";
-import { Translator } from "./types";
-
+export { default as constants } from "./constants";
 export { createTranslator } from './convertor'
-
-let quickTranslate: Translator;
-
-export const generate = () => {
-  if (quickTranslate === undefined) {
-    quickTranslate = createTranslator();
-  }
-  return quickTranslate.generate();
-}
-
-export const constants = alphabets;
-
-export default createTranslator;
-
+export { generate } from './quickTranslate'
