+++ src/quickTranslate.ts
@@ -1,0 +1,12 @@
+import { createTranslator } from "./convertor";
+
+import { Translator } from "./types";
+
+let quickTranslate: Translator;
+
+export const generate = () => {
+  if (quickTranslate === undefined) {
+    quickTranslate = createTranslator();
+  }
+  return quickTranslate.generate();
+}
